Create blocks table test helper module
